requirements.txt
torch
torchvision
mlflow
PyYAML
scikit-image
numpy
streamlit
plotly